Can now call virtual methods
We now write out the vtable ("isa") of each class and call init and dealloc through this vtable. We also init the vtables of subclasses with a copy of the superclass's vtable. We generate a header and source file for the base "object" class everything derives from. Checking in generated source code for example.ok so it's obvious what the output will be like.